refactor: Separate SpeechViewModel and timer move global thread
